+++ src/components/pages/apps/matcalc/matcalc-page.tsx
@@ -177,7 +177,7 @@ export function MatcalcPage() {
 
   const { setAppInfo } = useAppInfo()
 
-  const { settings, updateSettings } = useMatcalcSettings()
+  const { hasHydrated, settings, updateSettings } = useMatcalcSettings()
 
   const { sendMessage } = useMessages(MESSAGE_CHANNEL)
 
@@ -248,14 +248,19 @@ export function MatcalcPage() {
   }, [setToolbarTabs])
 
   useEffect(() => {
-    if (currentSelection?.id) {
-      updateSettings(
-        produce(settings, (draft) => {
-          draft.view.panels.tab = currentSelection.id
-        })
-      )
+    if (!hasHydrated || !currentSelection?.id) {
+      return
     }
-  }, [currentSelection])
+
+    if (settings.view.panels.tab === currentSelection.id) {
+      return
+    }
+    updateSettings(
+      produce(settings, (draft) => {
+        draft.view.panels.tab = currentSelection.id
+      })
+    )
+  }, [currentSelection, hasHydrated, settings, updateSettings])
 
   async function loadZTestData() {
     let res = await httpFetch.getText('/data/test/z_table.txt')